updates view titles
